app/requirements.txt
fastapi
uvicorn
transformers
safetensors
kafka-python
pydantic
torch
openpyxl
python-dateutil
psycopg2-binary
sqlalchemy
python-multipart
